handler: send persistant data while pulling/building images
when master becomes slave or nothing, we can keep going the service
while sending the first time data. in the same time
future master is pulling and building images.

What missing is that the old master should wait images are ready
before shutdonw and do the last snapshot, this will be the next
improvement